app/bot/requirements.txt
discord.py
pypresence
